simplicity: defend step-6 host complexity budget
- move RUB-598 host/context helpers out of program.go

- keep the step-6 evaluation invariant unchanged

- keep the Go diff within the RUB-598 LOC budget
